Implement Zobrist hashing (closes #55, fixes #3)
The previous position hasher was *not* accounting for

* en-passant,
* checks given in three-check and
* crazyhouse data.

It was therefore incorrectly giving some threefold repetitions.
This is now fixed.

In order to account for *castling*, the previous implementation was
clearing the transposition table on castling moves. Since the
transposition table is also used for the 50-move rule, this incorrectly
reset the half move counter (issue #3). This is now fixed.

In order to update the hash after a move (with castling rights and
en-passant privileges), the hasher now needs to work on a fully
finialized situation.

Zobrist hashing is based on a set of pre-agreed random bitmasks. The
particular ones chosen are compatible with the Polyglot opening book
format. This allows testing against given reference values. (Additionally
Polyglot opening book reading would now be easy to implement, but we
don't really have a use case for this, at the moment.)