Remove Ply from PgnNodeEncoder
If the We have the tree and a starting ply, then We don't really need Ply
for each pgn node. We can increment ply when traversing the tree. This
will reduce the complexity for generating PGN from other structure as We
don't have to increment ply when generating PgnNodeEncoder. This also
give us an option to not include turn number in pgn.