Fix deadlock class initialization
We started to use `Standard` more when initialize other object
especially with `Standard.initialPosition`. This increases the chances
of deadlock when initializing classes. Locally it happen 50% of times I
run test. This commit is to reducing computation time when initialize
Standard, and also reducing reference of Standard from other variants.
After this commits I'm not experienced deadlock anymore (with more than
10 tries)

- Use hard-coded initial Board for standard position
- Move initialPieces to def